pyproject.toml
[tool.poetry]
name = "cosmic"
version = "0.1.0"
description = ""
authors = ["Maksym Shalenyi <maksym.shalenyi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
mypy = "^0.770"
sqlalchemy = "^1.3.15"
"zope.interface" = "^5.0.1"
mypy-zope = "^0.2.5"
pydantic = "^1.7.3"
pip = "^21.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pdbpp = "^0.10.2"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
